pyproject.toml
[project]
name = "career-copilot"
version = "0.1.0"
description = "AI-powered job matching system with RAG"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Tirth Patel", email = "dev.tirthp14@gmail.com" }
]

dependencies = [
    # Resume file parsing
    "pypdf2==3.0.1",
    "python-docx==1.1.0",

    # Configuration
    "pyyaml==6.0.1",
    "python-dotenv==1.0.1",

    # Database
    "sqlalchemy==2.0.35",

    # LLM framework
    "langchain==0.3.1",
    "langchain-community==0.3.1",

    # Job scraping
    "beautifulsoup4==4.14.2",
    "requests>=2.31.0",

    # Vector database
    "chromadb>=0.4.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# Ruff configuration (fast Python linter)
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
]
ignore = []

# Black configuration (code formatter)
[tool.black]
line-length = 100
target-version = ["py310"]

# MyPy configuration (type checker)
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
